+++ client/draft/address.js
@@ -1,5 +1,6 @@
 import {getGlobalData} from "../src/utils/global_data";
 import {findMany} from "../src/utils/crud";
+import Taro from "@tarojs/taro";
 
 getAddressData = () => {
   let user_id = getGlobalData("user_id")
@@ -11,4 +12,25 @@ getAddressData = () => {
       shoppingAddress: res
     });
   })
+}
+
+
+// 获取用户收货地址，缓存无则重新获取
+getUserAddressData = () => {
+  let selectAddress = Taro.getStorageSync('ordersAddress')
+  if(selectAddress){
+    this.setState({
+      selectAddress
+    })
+  }else {
+    let user_id = getGlobalData("user_id")
+    let fields = ["id", "default", "username", "telephone", "province", "area", "city", "address"]
+    findMany({collection:"userAddress",condition:{user_id: user_id,default:1},fields}).then(res =>{
+      // console.log('orders userAddressData',res)
+      this.setState({
+        selectAddress: res[0]
+      })
+      Taro.setStorage({ key: 'ordersAddress', data: res[0] })
+    })
+  }
 }
